server/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
attrs==23.2.0
azure-core==1.30.2
azure-cosmos==4.7.0
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.6
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.0
executing==2.0.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
iniconfig==2.0.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.3
langchain-mongodb==0.1.6
langchain-openai==0.1.8
langchain-text-splitters==0.2.0
langsmith==0.1.68
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib-inline==0.1.7
mongomock==4.1.2
multidict==6.0.5
mypy-extensions==1.0.0
nest_asyncio==1.6.0
numpy==1.26.4
openai==1.31.0
orjson==3.10.3
packaging==23.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.2.2
pluggy==1.5.0
prompt_toolkit==3.0.46
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.7.3
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.8.0
pymongo==4.7.2
pypdf==4.2.0
pytest==8.2.1
pytest-env==1.1.3
python-dateutil==2.9.0
python-dotenv==1.0.1
PyYAML==6.0.1
pyzmq==26.0.3
regex==2024.5.15
requests==2.32.3
sentinels==1.0.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
stack-data==0.6.2
tavily-python==0.3.3
tenacity==8.3.0
tiktoken==0.7.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.3
yarl==1.9.4
zipp==3.17.0
